back-end

+++ src/main/java/com/toy/project/repository/CustomerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.toy.project.vo.CustomerVO;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<CustomerVO, Integer>{
+public interface CustomerRepository extends JpaRepository<CustomerVO, Long>{
 	
 }
